NP-16765 Move error logic
